feat(a11y): add dynamic screen reader summaries for chart accessibility
