+++ packages/charts/src/components/accessibility/screen_reader_summary.tsx
@@ -60,6 +60,5 @@ const ScreenReaderSummaryComponent = ({
   );
 };
 
-
 /** @internal */
 export const ScreenReaderSummary = memo(connect(mapStateToProps)(ScreenReaderSummaryComponent));
